requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
bs4==0.0.1
Django==1.10.5
djangorestframework==3.5.4
gunicorn==19.6.0
packaging==16.8
pyparsing==2.1.10
six==1.10.0
